Move field types into single file
