--- lib/geckoboard/date.rb
@@ -1,7 +1,0 @@
-module Geckoboard
-  class Date < Field
-    def to_hash
-      super.merge(type: :date)
-    end
-  end
-end
