--- lib/geckoboard/date_time.rb
@@ -1,7 +1,0 @@
-module Geckoboard
-  class DateTime < Field
-    def to_hash
-      super.merge(type: :datetime)
-    end
-  end
-end
