--- lib/geckoboard/number.rb
@@ -1,7 +1,0 @@
-module Geckoboard
-  class Number < Field
-    def to_hash
-      super.merge(type: :number)
-    end
-  end
-end
