--- lib/geckoboard/percentage.rb
@@ -1,7 +1,0 @@
-module Geckoboard
-  class Percentage < Field
-    def to_hash
-      super.merge(type: :percentage)
-    end
-  end
-end
