--- lib/geckoboard/string.rb
@@ -1,7 +1,0 @@
-module Geckoboard
-  class String < Field
-    def to_hash
-      super.merge(type: :string)
-    end
-  end
-end
